RAB-982: ServiceApi to fetch MeasurementFamilies
