+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/config/public_api.yml
@@ -14,7 +14,7 @@ services:
       arguments:
         - '@database_connection'
 
-    akeneo_measurement.public_api.onboarder.find_all_measurement_families:
-        class: 'Akeneo\Tool\Bundle\MeasureBundle\PublicApi\Onboarder\FindAllMeasurementFamilies'
+    akeneo_measurement.public_api.onboarder.find_measurement_families:
+        class: 'Akeneo\Tool\Bundle\MeasureBundle\PublicApi\FindMeasurementFamilies'
         arguments:
             - '@akeneo_measure.persistence.measurement_family_repository'
